添加登录身份功能

+++ controllers/todo/removeTodos.js
@@ -1,24 +1,29 @@
-function remove(model,todoId) {
-    return model.findOneAndRemove({id:todoId});
+function remove(model,todoId,userId) {
+    return model.findOneAndRemove({id:todoId,userId});
 }
 
 module.exports = async (ctx, next) => {
-    let {idList} = ctx.query;
-    console.log(idList);
-    if(!idList || !idList.length){
-        ctx.error("id不能为空");
-        return;
-    }
-    const Todo = require('../../models/Todo.js');
-    idList = idList.split(',');
-    try {
-        idList.map(async (item)=>{
-            let data = await remove(Todo,item);
-            console.log(data);
-        });
-        ctx.success();
-    } catch (e) {
-        ctx.error(e);
+    const requestData = ctx.method === 'POST' ? ctx.request.body : ctx.query;
+    let {idList} = requestData;
+    const {userId} = ctx.session;
+    if(ctx.session.userId){
+        if(!idList || !idList.length){
+            ctx.error("id不能为空");
+            return;
+        }
+        const Todo = require('../../models/Todo.js');
+        idList = idList.split(',');
+        try {
+            idList.map(async (item)=>{
+                let data = await remove(Todo,item,userId);
+                console.log(data);
+            });
+            ctx.success();
+        } catch (e) {
+            ctx.error(e);
+        }
+    }else {
+        ctx.error("登录身份过期，请重新登录");
     }
     await next();
 };
